# api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "A high-performance email verification system with microservices architecture\n\nThis API provides comprehensive email verification services including:\n- Syntax validation and role-based detection\n- Domain validation with DNS and MX record checking\n- SMTP server connectivity and capability detection\n- Inbox existence verification\n- Bulk email verification with task management\n- Proxy management for distributed verification\n- Role pattern and disposable domain management",
        "title": "Email Verification API",
        "contact": {
            "name": "API Support",
            "url": "http://example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/emails": {
            "get": {
                "description": "Get a paginated list of all verified emails with their verification results",
                "tags": [
                    "emails"
                ],
                "summary": "List emails",
                "parameters": [
                    {
                        "description": "Number of emails to return per page",
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 100,
                            "default": 10
                        }
                    },
                    {
                        "description": "Number of emails to skip for pagination",
                        "name": "offset",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of emails",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/cmd_controllers.EmailListResponse"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
        "/tasks": {
            "get": {
                "description": "Get a paginated list of verification tasks with optional filtering by user ID and creation date range",
                "tags": [
                    "tasks"
                ],
                "summary": "List tasks",
                "parameters": [
                    {
                        "description": "Number of tasks to return per page",
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 100,
                            "default": 10
                        }
                    },
                    {
                        "description": "Number of tasks to skip for pagination",
                        "name": "offset",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0
                        }
                    },
                    {
                        "description": "Filter by user ID (UUID format)",
                        "name": "user_id",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "description": "Filter tasks created after this timestamp (RFC3339 format, e.g., 2023-01-01T00:00:00Z)",
                        "name": "created_after",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "description": "Filter tasks created before this timestamp (RFC3339 format, e.g., 2023-12-31T23:59:59Z)",
                        "name": "created_before",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/cmd_controllers.TaskListResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters (malformed UUID or timestamp format)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/validation.ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "description": "Creates a new task for bulk email verification with a maximum of 10,000 emails per task. Emails are grouped by domain and queued for processing. Optionally configure a webhook URL to receive notifications when verification completes.",
                "tags": [
                    "tasks"
                ],
                "summary": "Create a bulk verification task",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/cmd_controllers.TaskRequest"
                            }
                        }
                    },
                    "description": "Task creation request with emails, optional user ID, and optional webhook URL",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/cmd_controllers.TaskResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request (exceeds 10,000 email limit, invalid emails, malformed UUID, invalid webhook URL)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/validation.ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error during task creation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/tasks/batch/upload": {
            "post": {
                "description": "Upload a file containing email addresses for batch verification. File size limit and allowed formats are determined by server configuration. Returns immediately with upload confirmation while processing happens in background.",
                "tags": [
                    "tasks"
                ],
                "summary": "Upload file for batch email verification",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "description": "Email list file",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "webhook_url": {
                                        "description": "Webhook URL for notifications",
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "description": "User ID (UUID format)",
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "file"
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "File uploaded and processing started",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/cmd_controllers.BatchFileUploadResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid file or validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/cmd_controllers.BatchFileUploadError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
        "/tasks/{id}": {
            "get": {
                "description": "Get detailed information about a specific task including all email verification job results with email, host, and domain relationships",
                "tags": [
                    "tasks"
                ],
                "summary": "Get task details",
                "parameters": [
                    {
                        "description": "Task ID (UUID format)",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task details with email results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/cmd_controllers.TaskDetailResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid task ID format",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/validation.ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "description": "API key authorization header (Bearer token)",
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
...
            "cmd_controllers.BatchFileUploadError": {
                "type": "object",
                "properties": {
                    "details": {
                        "type": "string"
                    },
                    "error": {
                        "type": "string"
                    }
                }
            },
...
            "cmd_controllers.EmailListResponse": {
                "type": "object",
                "properties": {
                    "count": {
                        "type": "integer",
                        "example": 1000
                    },
                    "emails": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/models.Email"
                        }
                    },
                    "limit": {
                        "type": "integer",
                        "example": 10
                    },
                    "offset": {
                        "type": "integer",
                        "example": 0
                    }
                }
            },
...
            "cmd_controllers.TaskListResponse": {
                "type": "object",
                "properties": {
                    "count": {
                        "type": "integer",
                        "example": 150
                    },
                    "limit": {
                        "type": "integer",
                        "example": 10
                    },
                    "offset": {
                        "type": "integer",
                        "example": 0
                    },
                    "tasks": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/models.Task"
                        }
                    }
                }
            },
            "cmd_controllers.TaskRequest": {
                "type": "object",
                "required": [
                    "emails"
                ],
                "properties": {
                    "emails": {
                        "type": "array",
                        "maxItems": 10000,
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "[\"user1@example.com\"",
                            " \"user2@example.com\"]"
                        ]
                    },
                    "user_id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "webhook_url": {
                        "type": "string",
                        "example": "https://example.com/webhook"
                    }
                }
            },
...
            "validation.ValidationErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    },
                    "fields": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}